Retargeting: refactor client side redirect to API route
